separated merger into header and body
